remove file with just a helper function and make test pass
